test(cli): extract runCli, writeResultArtifact, makeTmpDir to test/_helpers.js
Move the three load-bearing CLI test primitives into a shared helper
module so a second test file can consume them without copy-divergence.
`makeTmpDir` returns { dir, cleanup } so each test keeps its explicit
try/finally lifecycle — no hidden global test state.

What stays inline: binPath-style one-liners and test-specific fixture
constants that have a single call site today. Extract those only when a
second consumer arrives.
